refactor: rename `constants` package to `osvschema`
